- renames namespaces from app to App - refactors to short closures - refactors collect into Collection - adds typed properties in services, where possible - adds return types in service methods - uses permission Types enum in structure migration - adds missing facade imports - addresses phpinsights - removes resource - moves Api to services
