+++ composer.json
@@ -3,8 +3,8 @@
     "description": "A control panel that aggregates, controls and displays statistics from other laravel-enso apps",
     "type": "project",
     "require": {
-        "laravel-enso/core": "4.6.*",
-        "laravel-enso/enums": "1.1.*",
+        "laravel-enso/core": "4.7.*",
+        "laravel-enso/enums": "1.2.*",
         "guzzlehttp/guzzle": "^6.3"
     },
     "license": "MIT",
