+++ src/app/Http/Controllers/Application/ExportExcel.php
@@ -4,7 +4,7 @@ namespace LaravelEnso\ControlPanel\app\Http\Controllers\Application;
 
 use Illuminate\Routing\Controller;
 use LaravelEnso\ControlPanel\app\Tables\Builders\ApplicationTable;
-use LaravelEnso\Tables\app\Traits\Excel;
+use LaravelEnso\Tables\App\Traits\Excel;
 
 class ExportExcel extends Controller
 {
