+++ src/app/Http/Controllers/Application/InitTable.php
@@ -4,7 +4,7 @@ namespace LaravelEnso\ControlPanel\app\Http\Controllers\Application;
 
 use Illuminate\Routing\Controller;
 use LaravelEnso\ControlPanel\app\Tables\Builders\ApplicationTable;
-use LaravelEnso\Tables\app\Traits\Init;
+use LaravelEnso\Tables\App\Traits\Init;
 
 class InitTable extends Controller
 {
