+++ src/app/Http/Controllers/Application/TableData.php
@@ -4,7 +4,7 @@ namespace LaravelEnso\ControlPanel\app\Http\Controllers\Application;
 
 use Illuminate\Routing\Controller;
 use LaravelEnso\ControlPanel\app\Tables\Builders\ApplicationTable;
-use LaravelEnso\Tables\app\Traits\Data;
+use LaravelEnso\Tables\App\Traits\Data;
 
 class TableData extends Controller
 {
