+++ src/app/Http/Controllers/Application/Update.php
@@ -3,7 +3,7 @@
 namespace LaravelEnso\ControlPanel\app\Http\Controllers\Application;
 
 use Illuminate\Routing\Controller;
-use LaravelEnso\ControlPanel\app\Http\Requests\ValidateApplicationRequest;
+use LaravelEnso\ControlPanel\App\Http\Requests\ValidateApplicationRequest;
 use LaravelEnso\ControlPanel\app\Models\Application;
 
 class Update extends Controller
