+++ src/app/Http/Responses/ClearLogResponse.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace LaravelEnso\ControlPanel\app\Http\Responses;
+namespace LaravelEnso\ControlPanel\App\Http\Responses;
 
 class ClearLogResponse extends ApiResponse
 {
-    public function method()
+    public function method(): string
     {
         return 'clearLog';
     }
