+++ src/routes/app/applications.php
@@ -1,5 +1,7 @@
 <?php
 
+use Illuminate\Support\Facades\Route;
+
 Route::prefix('administration')
     ->as('administration.')
     ->group(function () {
