+++ src/routes/app/controlPanel.php
@@ -1,5 +1,7 @@
 <?php
 
+use Illuminate\Support\Facades\Route;
+
 Route::namespace('ControlPanel')
     ->prefix('controlPanel')
     ->as('controlPanel.')
